Move Arrow component to new sub component
