+++ assets/js/components/UnnumberedPagination/Arrow.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+
+function Arrow({ children, onClick }) {
+  return (
+    <div
+      aria-hidden="true"
+      role="button"
+      className="cursor-pointer"
+      onClick={() => onClick()}
+    >
+      {children}
+    </div>
+  );
+}
+
+export default Arrow;
